fix: improve benchmark presentation
